Moved mailroom-nest implementation from app to lib
