+++ apps/mailroom-nest/src/app/app.module.ts
@@ -2,17 +2,16 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { MailroomEmail, MailroomAttachment, MailroomReject } from '@tamu-gisc/mailroom/common';
+import { EmailModule } from '@tamu-gisc/mailroom/data-api';
 
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { DB_CONFIG } from '../environments/environment';
 
 @Module({
   imports: [
     TypeOrmModule.forRoot({ ...DB_CONFIG, entities: [MailroomEmail, MailroomAttachment, MailroomReject] }),
-    TypeOrmModule.forFeature([MailroomEmail, MailroomAttachment, MailroomReject])
+    EmailModule
   ],
-  controllers: [AppController],
-  providers: [AppService]
+  controllers: [],
+  providers: []
 })
 export class AppModule {}
