+++ libs/mailroom/data-api/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/email/email.module';
