+++ libs/mailroom/data-api/src/lib/email/email.module.ts
@@ -1,10 +1,14 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { MailroomEmail, MailroomAttachment, MailroomReject } from '@tamu-gisc/mailroom/common';
+
+import { EmailController } from './email.controller';
 import { EmailService } from './email.service';
 
 @Module({
-  imports: [],
-  controllers: [],
+  imports: [TypeOrmModule.forFeature([MailroomEmail, MailroomAttachment, MailroomReject])],
+  controllers: [EmailController],
   providers: [EmailService],
   exports: []
 })
